Create an element from a ComponentClass
An example use-case for this function is when a component class is
imported and it is desired to render an instance of this class within
another component or as the top-level component.

For instance, rackt/react-router exports Route, DefaultRoute,
RouteHandler, etc. The createElementFromClass function can be used to
render these exported classes.