change the way React.createElement is called
this is a change for paf31/purescript-thermite#27

when you return an array of JSX elements, it passes
an array as children in createElement, so this works:

    <div>
      { this.props.map(function(el) {
        return <span>{el}</span>;
      }}
    </div>

but when you structure your tree like this:

    <div>
      <span>one</span>
      <span>two</span>
    </div>

it calls createElement(..., <span>one</span>, <span>two</span>).

this is important because elements passed in the 2nd way
don't require unique keys and reacts fills them in automatically.

when passing elements the first way you always have to give unique
keys yourself